Handle all enum integer value types

+++ src/ProtoRebuilder/Models/EnumInfo.cs
@@ -1,5 +1,6 @@
 ﻿using Mono.Cecil;
 using System.Diagnostics;
+using System.Numerics;
 
 namespace Knapcode.ProtoRebuilder;
 
@@ -9,5 +10,5 @@ public class EnumInfo
     public bool IsRoot => RootMessage is null;
     public required MessageInfo? RootMessage { get; init; }
     public required TypeDefinition Type { get; init; }
-    public required IReadOnlyList<KeyValuePair<string, int>> Pairs { get; init; }
+    public required IReadOnlyList<KeyValuePair<string, BigInteger>> Pairs { get; init; }
 }
